+++ packages/examples/ts-server/src/controllers/user.ts
@@ -5,16 +5,23 @@ import {
   ReqParams,
   Post,
   ReqBody,
-  FWResponse,
-  FWRequest,
-  Response
+  InjectCtx,
+  Inject,
+  Context
 } from 'flash-wolves'
+import TestService from '../services/Test'
 
 @RouterController('user', { requiresAuth: true })
 export default class User {
   private hello() {
-    console.log('hello')
+    console.log('hello', this.ctx.req.url)
   }
+
+  @Inject(TestService)
+  private test: TestService
+
+  @InjectCtx()
+  private ctx: Context
 
   @Get('info/:id', { isAdmin: true })
   getUserInfo(
@@ -25,11 +32,17 @@ export default class User {
     // req: FWRequest,
     // res: FWResponse
   ) {
+    this.test.hello()
     this.hello()
+    console.log('userInfo', this.ctx.userInfo)
+
     console.log(query, params, id, search)
     // res.plain('<h1>Hello World</h1>', 'text/html')
     // return Response.plain('123')
-    return query
+    return {
+      query,
+      params
+    }
   }
 
   @Post('login')